added weightclass struct and way to assign weightclasses to players

+++ Assets/Scripts/Combat/AttackHurtbox.cs
@@ -3,9 +3,9 @@ using UnityEngine;
 public class AttackHurtbox : MonoBehaviour
 {
     public Player player;
-    public int HP = 200;
+    public float HP = 200;
 
-    public void TakeDamage(int damage)
+    public void TakeDamage(float damage)
     {
         if (player != null)
             player.TakeDamage(damage);
